Moved actions, mutations, and getters to separate files
