+++ molgenis-questionnaires/src/main/frontend/src/store/getters.js
@@ -1,0 +1,11 @@
+const getters = {
+  getChapterByIndex: (state) => (index) => {
+    return [state.chapterFields[(index - 1)]]
+  },
+
+  getTotalNumberOfChapters: state => {
+    return state.chapterFields.length
+  }
+}
+
+export default getters
